Embed copy of cram
We use the cram Python module [1] for command-line testing — i.e., where
you want to run a command line program and verify its output in some
way.  Before we relied on the cram module being installed in some system
location.  But it's not very big, and it's only a single file, so lets
embed it directly in the source tree instead.  Now you only need to have
Python installed to be able to run the command-line tests.

[1] https://pypi.python.org/pypi/cram