Move Sphinx documentation
I'm about to merge in the latest C template, which switches the
documentation from Sphinx to man pages.  I don't want to completely
delete the Sphinx documentation, at least not until it's been translated
to some other format.  So this patch moves it into a different location,
so that we can build both man pages and Sphinx for now.