Fix IPv6 parse error stack overflow
Mac OS X has some better stack overflow detection, and it caught a
problem with the IPv6 parser.  A particular kind of parse error could
cause us to write an extra hextet into our temporary buffer, which we
hadn't allocated room for.  This patch adds an extra check to catch this
situation.