Don't use mempool in cork_hash_table
Before we were using a `cork_mempool` to allocate all of the entries
that are stored in a hash table.  This reduces the number of times that
we have to call `malloc`, but currently the mempool implementation isn't
able to return memory to the OS.  That makes it less than ideal for a
hash table that grows and shrinks during its existence.

Plus, now that we have customizable allocation, a better approach is for
your application to use a better `malloc` implementation, like jemalloc
or tcmalloc.