Separate out major/minor/patch versions
Our template already grabs the current version from git tags.  We now
also separate this out in major, minor, and patch versions, stored in
the `VERSION_MAJOR`, `VERSION_MINOR`, and `VERSION_PATCH` cmake
variables, respectively.  We also grab the SHA-1 identifier of the
current git commit and store that in the `GIT_SHA1` variable.

(Note that this means that the cmake scripts only work from a valid git
checkout.  The version detection tries to support working from a release
tarball, as well, but we don't currently have a good mechanism for
generating one of those.  Once we do, we should tweak all of this to
work there, too.)