Fix macros for cmake 2.8.7
We need to be careful about when we're passing the `__SOURCES` variable
as a list (and therefore don't want to quote it).  We can't use
`target_include_directories` before 2.8.12, and need to fall back on
`include_directories` if it isn't available.  And we have to explicitly
include the file that defines `cmake_parse_arguments` on some earlier
cmake versions (although I'm not sure precisely where the boundary is).