Debug allocator should always abort
We provide a "debug" allocator, which verifies that the sizes that you
provide to the allocation functions matches the sizes that you provide
to the deallocation functions.  Before, we used the builtin `assert`
macro to perform this check.  However, that means that if we compile
libcork in `Release` mode (in which asserts are left out), we never
actually perform the check!  We now use an explicit check, and call
`abort` ourselves if the check fails — our assumption being that if you
ask for the debug allocator, you want at least this particular bit of
debug behavior even if the rest of the library was compiled in `Release`
mode.