Fix the build on i386
Our Travis Linux builds run on a 64-bit machine, even when we're
compiling and testing in 32-bit mode.  That means pkg-config by default
will look in the 64-bit arch-specific directory for libcheck, even
though we were (correctly) installing the 32-bit version of the library.
Debian ships a helper script for running pkg-config in cross-compile
mode (/usr/share/pkg-config-crosswrapper), but all that does is set
PKG_CONFIG_PATH to the appropriate values.  We can do that directly in
our Travis script.