Fix processing of libtool-style version-info
We were calculating the SONAME correctly, but not the version encoded
into the .so filename.  This patch updates our CMake rules to line up
with what libtool does, and changes the name and format of the paramater
to make it clearer that this should line up with libtool's -version-info
parameter.