Use Travis container infrastructure
This should speed up our builds.  The only wrinkle is that we have a
package that we only technically need when performing a 32-bit build.
There's no way (that I know of) to conditionally install that package in
the container world like we were doing before.  But it doesn't hurt
anything to install the package when it's not needed.