Add CORK_INLINE attribute (#161)
Ideally, we use the C99 `inline` attribute — you provide a declaration
_and_ definition of your function in a header file, labeled with the new
CORK_INLINE attribute.  You then provide a copy of the declaration (but
_not_ the definition) in a source file, _without_ the CORK_INLINE
attribute.  That ensures that there's a linkable copy of the function as
part of that source file — and that its definition is guaranteed to
match the one in the header file.

If C99 inline isn't available, we fall back on using `static`.  That
means that the function definition will be duplicated in each source
file that uses it, but prevents duplicate symbol definition errors.
(That means you can't rely on being able to take the address of the
inline function, in a way that's consistent across source files, unless
you can guarantee that you're compiling in C99 mode.)

This patch also updates a lot of libcork functions and macros to use the
new attribute.